Create ideologies.json
Japanese translation. I wanted to add the original word in brackets (e.g. "ideo1": "無政府共産主義（Anarcho-Capitalism）"), but it would be too long and would be very ugly without significant changes in the CSS and the banner script (especially with "ideo7"). Clarifying the original name is important in my opinion, since many of these ideologies are not widely known in Japan, and sometimes even googling in Japanese yields almost zero results . LeftValues, which I also contributed to, avoids this problem by adding a description section, and I think implementing a similar solution in the future is desirable.